requirements.txt
anyio==4.1.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.42.0
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
DataProperty==1.0.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
dominate==2.9.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.46.0
fqdn==1.5.1
fsspec==2023.10.0
ftfy==6.1.3
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.41
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
loguru==0.7.2
lpips==0.1.4
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.4
objaverse==0.1.7
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathvalidate==3.2.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
piqa==1.3.2
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
prettytable==3.9.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.2
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pycparser==2.21
Pygments==2.17.2
PyMCubes==0.1.4
pyparsing==3.1.1
pypistats==1.5.0
PySocks==1.7.1
pytablewriter==1.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.1
safetensors==0.4.1
scipy==1.10.1
Send2Trash==1.8.2
sentry-sdk==1.40.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tabledata==1.3.3
tcolorpy==0.1.4
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.35.2
typepy==1.3.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wandb==0.16.2
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
zipp==3.17.0
